ai/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai-whisper==20231117
yt-dlp==2023.11.16
python-multipart==0.0.6
ffmpeg-python==0.2.0

# For deployment
gunicorn==21.2.0

